Refactor current backend code into oldapi, create new components at /api/ instead of /rest/api
