+++ backend/app.js
@@ -5,14 +5,18 @@ const cors = require('cors');
 const path = require('path');
 const mongoose = require("mongoose");
 const cookieParser = require('cookie-parser');
+
 const config = require('./utils/config');
 const logger = require('./utils/logger');
 const middleware = require('./utils/middleware')
+
 const indexRouter = require('./routes/index');
 const statusRouter = require('./routes/status');
+const oldapiRouter = require('./controllers/oldapi');
 const usersRouter = require('./controllers/users');
+const loginRouter = require('./controllers/login');
 
-var app = express();
+const app = express();
 
 // Attach middlewares
 app.use(cors());
@@ -35,10 +39,14 @@ app.use('/', indexRouter);
 app.use('/status', statusRouter);
 
 // API Routes
-app.use('/rest/api', usersRouter);
+app.use('/rest/api', oldapiRouter);
+app.use('/api/users', usersRouter);
+app.use('/api/login', loginRouter);
 
 // Custom processing of errorMessage (mostly related to mongoose)
+// Using `express-async-errors` package, no need to wrap async calls with try/catch, but need middleware
+// This middleware wraps async calls automatically and does errorHandling when error is thrown
 // Must be last loaded middleware
-
 app.use(middleware.errorHandler)
+
 module.exports = app;
