+++ backend/controllers/login.js
@@ -1,0 +1,4 @@
+const loginRouter = require('express').Router()
+const User = require('../models/user')
+
+module.exports = loginRouter
